@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # == Schema Information
 #
 # Table name: balances
@@ -10,11 +12,11 @@
 #  team_id    :integer
 #
 
-class Balance < ActiveRecord::Base
+class Balance < ApplicationRecord
   validates :name, presence: true
 
   belongs_to :team
-  has_many :transactions
+  has_many :posts
   has_many :goals
 
   attr_accessor :make_balance_active_checkbox
@@ -25,16 +27,16 @@ class Balance < ActiveRecord::Base
     where(current: true).where(team_id: team).order("created_at DESC").first
   end
 
-  def last_transaction
-    transactions.order("created_at DESC").first
+  def last_post
+    posts.order("created_at DESC").first
   end
 
   def self.likes(balance)
-    Transaction.joins("INNER JOIN votes on votes.votable_id = transactions.id and votable_type='Transaction'").where("balance_id=#{balance.id}").count
+    Post.joins("INNER JOIN votes on votes.votable_id = posts.id and votable_type='Post'").where("balance_id=#{balance.id}").count
   end
 
   def amount
-    Transaction.where(balance: self).sum(:amount) + Balance.likes(self)
+    Post.where(balance: self).sum(:amount) + Balance.likes(self)
   end
 
   def self.time_left
@@ -43,6 +45,4 @@ class Balance < ActiveRecord::Base
     days_left = (expire - current).to_i / (24 * 60 * 60)
     "#{days_left} days left"
   end
-
-
 end
