@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # == Schema Information
 #
 # Table name: guidelines
@@ -10,9 +12,9 @@
 #  team_id    :integer
 #
 
-class Guideline < ActiveRecord::Base
+class Guideline < ApplicationRecord
   validates :name, presence: true, length: { minimum: 4, maximum: 100 }
-  validates :kudos, presence: true, numericality: { greater_than_or_equal_to: 1, less_than_or_equal_to: 1000, message: "is not correct. You can't give negative ₭udos or exceed over 1000"}
+  validates :kudos, presence: true, numericality: { greater_than_or_equal_to: 1, less_than_or_equal_to: 1000, message: "is not correct. You can't give negative ₭udos or exceed over 1000" }
 
   belongs_to :team
 
