@@ -40,7 +40,10 @@
 #
 
 
-class User < ActiveRecord::Base
+class User < ApplicationRecord
+  validates :name, presence: true
+  validates :email, presence: true, uniqueness: true
+
   after_update :ensure_an_admin_remains
 
   acts_as_voter
@@ -51,16 +54,26 @@ class User < ActiveRecord::Base
   devise :omniauthable, :registerable, :confirmable, :database_authenticatable,
          :validatable, :lockable, :recoverable, omniauth_providers: [:slack]
 
-  has_many :sent_transactions, class_name: 'Transaction', foreign_key: :sender_id
-  has_many :received_transactions, class_name: 'Transaction', foreign_key: :receiver_id
-  has_many :memberships, class_name: 'TeamMember', foreign_key: :user_id
+  has_many :sent_posts,
+           class_name: "Post",
+           foreign_key: "sender_id",
+           dependent: :destroy
+
+  has_many :post_receivers,
+           dependent: :destroy
+  has_many :received_posts,
+           through: :post_receivers,
+           source: :post
+
+
+  has_many :memberships, class_name: "TeamMember", foreign_key: :user_id
   has_many :teams, through: :memberships
-  has_many :votes, foreign_key: 'voter_id'
-  has_many :exports, foreign_key: 'user_id'
+  has_many :votes, foreign_key: "voter_id"
+  has_many :exports, foreign_key: "user_id"
   has_many :fcm_tokens
 
   typed_store :preferences, coder: PreferencesCoder do |p|
-    p.boolean :transaction_received_mail, default: true
+    p.boolean :post_received_mail, default: true
     p.boolean :goal_reached_mail, default: true
     p.boolean :summary_mail, default: true
   end
@@ -68,18 +81,18 @@ class User < ActiveRecord::Base
   def self.from_omniauth(access_token)
     data = access_token.info
 
-    email_address = data['email']
+    email_address = data["email"]
     email_domain = extract_email_domain(email_address)
     return if email_domain_not_allowed?(email_domain)
 
     existing_user = User.exists? uid: access_token.uid
 
     user = User.where(uid: access_token.uid).first_or_create(
-        provider: access_token.provider,
-        uid: access_token.uid,
-        name: data['name'],
-        email: email_address,
-        avatar_url: data['image']
+      provider: access_token.provider,
+      uid: access_token.uid,
+      name: data["name"],
+      email: email_address,
+      avatar_url: data["image"]
     )
 
     UserMailer.new_user(user) unless existing_user
@@ -87,33 +100,8 @@ class User < ActiveRecord::Base
     user
   end
 
-  def self.from_api_token_request(params)
-    email_domain = extract_email_domain(params['email'])
-    if email_domain_not_allowed?(email_domain)
-      error_object_overrides = {title: 'Invalid email domain', detail: "Email domain #{email_domain} is not allowed."}
-      error = Api::V1::UnauthorizedError.new(error_object_overrides)
-      raise error
-    end
-
-    existing_user = User.exists? uid: params['uid']
-
-    user = User.where(uid: params['uid']).first_or_create(params.except(:jwt_token))
-
-    if user.deactivated?
-      error_object_overrides = {title: 'Deactivated user account', detail: "User #{user.name} is deactivated."}
-      error = Api::V1::UnauthorizedError.new(error_object_overrides)
-      raise error
-    end
-
-    user.update(api_token: generate_unique_api_token) unless user.api_token.present?
-
-    UserMailer.new_user(user) unless existing_user
-
-    user
-  end
-
   def self.find_by_term(term)
-    User.order(:name).where('lower(name) like ?', "#{term}%".downcase)
+    User.order(:name).where("lower(name) like ?", "#{term}%".downcase)
         .where(deactivated_at: nil).where(restricted: false)
   end
 
@@ -137,20 +125,20 @@ class User < ActiveRecord::Base
     TeamInvite.where(email: self.email, accepted_at: nil, declined_at: nil)
   end
 
-  def all_transactions
-    Transaction.all_for_user(self)
+  def all_posts
+    Post.all_for_user(self)
   end
 
   def first_name
-    name.split(' ')[0]
+    name.split(" ")[0]
   end
 
   def picture_url
-    avatar_url.presence || '/no-picture-icon.jpg'
+    avatar_url.presence || "/no-picture-icon.jpg"
   end
 
   def deactivate
-    transaction do
+    post do
       update(deactivated_at: DateTime.now, api_token: nil)
       fcm_tokens.destroy_all
       ensure_an_admin_remains
@@ -199,24 +187,24 @@ class User < ActiveRecord::Base
 
   private
 
-  def ensure_an_admin_remains
-    if User.where(admin: true, deactivated_at: nil).count < 1
-      raise "Last administrator can't be removed from the system"
+    def ensure_an_admin_remains
+      if User.where(admin: true, deactivated_at: nil).count < 1
+        raise "Last administrator can't be removed from the system"
+      end
     end
-  end
 
-  def self.extract_email_domain(email_address)
-    email_address.split('@')[1]
-  end
+    def self.extract_email_domain(email_address)
+      email_address.split("@")[1]
+    end
 
-  def self.email_domain_not_allowed?(email_domain)
-    email_domain != ENV.fetch('DEVISE_DOMAIN', 'gmail.com')
-  end
+    def self.email_domain_not_allowed?(email_domain)
+      email_domain != ENV.fetch("DEVISE_DOMAIN", "gmail.com")
+    end
 
-  def self.generate_unique_api_token
-    api_token = SecureRandom.urlsafe_base64
-    # recursively call this function to ensure that a unique api token is generated
-    generate_unique_api_token if User.exists?(api_token: api_token)
-    api_token
-  end
+    def self.generate_unique_api_token
+      api_token = SecureRandom.urlsafe_base64
+      # recursively call this function to ensure that a unique api token is generated
+      generate_unique_api_token if User.exists?(api_token: api_token)
+      api_token
+    end
 end
