@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # == Schema Information
 #
 # Table name: votes
@@ -14,11 +16,11 @@
 #  updated_at   :datetime
 #
 
-class Vote < ActiveRecord::Base
-  belongs_to :user_voter, class_name: 'User', foreign_key: 'voter_id'
-  belongs_to :transaction_votable, class_name: 'Transaction', foreign_key: 'votable_id'
+class Vote < ApplicationRecord
+  belongs_to :user_voter, class_name: "User", foreign_key: "voter_id"
+  belongs_to :post_votable, class_name: "Post", foreign_key: "votable_id"
 
   def self.all_for_user(user)
-    Vote.where(voter_id: user.id).order('created_at desc')
+    Vote.where(voter_id: user.id).order("created_at desc")
   end
 end
