@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 # Load the Rails application.
-require_relative 'application'
+require_relative "application"
 
 # Initialize the Rails application.
 Rails.application.initialize!
 
-ActionMailer::Base.default :from => ENV['MAIL_USERNAME']
+ActionMailer::Base.default from: ENV["MAIL_USERNAME"]
 
 # Remove field_with_error divs
 ActionView::Base.field_error_proc = Proc.new do |html_tag, instance|
