@@ -1,4 +1,6 @@
-class AddAttachmentZipToExports < ActiveRecord::Migration
+# frozen_string_literal: true
+
+class AddAttachmentZipToExports < ActiveRecord::Migration[5.0]
   def self.up
     change_table :exports do |t|
       t.attachment :zip
