@@ -1,4 +1,6 @@
-require 'administrate/base_dashboard'
+# frozen_string_literal: true
+
+require "administrate/base_dashboard"
 
 class GoalDashboard < Administrate::BaseDashboard
   # ATTRIBUTE_TYPES
@@ -10,7 +12,7 @@ class GoalDashboard < Administrate::BaseDashboard
   ATTRIBUTE_TYPES = {
       id: Field::Number,
       name: Field::String,
-      amount: Field::Number.with_options(suffix: ' ₭'),
+      amount: Field::Number.with_options(suffix: " ₭"),
       achieved_on: Field::DateTime,
       created_at: Field::DateTime,
       updated_at: Field::DateTime,
