@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class GoalDecorator < Draper::Decorator
   delegate_all
 
@@ -8,5 +10,4 @@ class GoalDecorator < Draper::Decorator
   def kudos
     h.number_to_kudos(object.amount)
   end
-
 end
