@@ -1,4 +1,6 @@
-class TransactionDecorator < Draper::Decorator
+# frozen_string_literal: true
+
+class PostDecorator < Draper::Decorator
   delegate_all
   decorates_association :sender
   decorates_association :receiver
