@@ -1,15 +1,15 @@
-module ApplicationHelper
+# frozen_string_literal: true
 
+module ApplicationHelper
   def number_to_kudos(number)
     options = {
       precision: 0,
-      unit: '₭',
-      separator: ',',
-      delimiter: '.',
-      format: '%n %u',
+      unit: "₭",
+      separator: ",",
+      delimiter: ".",
+      format: "%n %u",
     }
 
     number_to_currency(number, options)
   end
-
 end
