@@ -1,12 +1,13 @@
+# frozen_string_literal: true
+
 class ApplicationMailer < ActionMailer::Base
-  default from: Rails.env == 'test' || ENV['MAIL_USERNAME'].blank? ? 'example@mail.com' : ENV['MAIL_USERNAME']
-  layout 'mailer'
-  add_template_helper(TransactionsHelper)
+  default from: Rails.env == "test" || ENV["MAIL_USERNAME"].blank? ? "example@mail.com" : ENV["MAIL_USERNAME"]
+  layout "mailer"
   before_action :add_logo_attachment
 
   private
 
-  def add_logo_attachment
-    attachments.inline['logo.png'] = File.read("#{Rails.root}/app/assets/images/kudo-o-matic-white-mail.png")
-  end
+    def add_logo_attachment
+      attachments.inline["logo.png"] = File.read("#{Rails.root}/app/assets/images/kudo-o-matic-white-mail.png")
+    end
 end
