@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 class GoalMailer < ApplicationMailer
   def self.new_goal(goal, team)
-    return if Rails.env == 'test' || ENV['MAIL_USERNAME'].blank?
+    return if Rails.env == "test" || ENV["MAIL_USERNAME"].blank?
 
-    team.users.where.not(email: '').where(deactivated_at: nil).each do |user|
-      suppress(Exception) {goal_email(user, team, goal).deliver_later if user.goal_reached_mail}
+    team.users.where.not(email: "").where(deactivated_at: nil).each do |user|
+      suppress(Exception) { goal_email(user, team, goal).deliver_later if user.goal_reached_mail }
     end
   end
 
@@ -11,6 +13,6 @@ class GoalMailer < ApplicationMailer
     @goal = goal
     @user = user
 
-    mail(to: user.email, subject: "Goal '#{Goal.previous(team).name}' is reached! \u{1f389}")
+    # mail(to: user.email, subject: "Goal '#{Goal.previous(team).name}' is reached! \u{1f389}")
   end
 end
