@@ -15,12 +15,12 @@ class TeamAdder
 
   private
 
-  def self.save(team, current_user)
-    if team.save!
-      team.add_member(current_user, true)
-      TransactionAdder.create_for_new_team(team, current_user)
-    end
+    def self.save(team, current_user)
+      if team.save!
+        team.add_member(current_user, true)
+        PostAdder.create_for_new_team(team, current_user)
+      end
 
-    team
-  end
+      team
+    end
 end
