@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
@@ -12,7 +14,7 @@ Rails.application.configure do
           access_key_id: ENV["AWS_ACCESS_KEY_ID"],
           secret_access_key: ENV["AWS_SECRET_ACCESS_KEY"],
           path: "image/:id/:filename",
-          url: ':s3_domain_url'
+          url: ":s3_domain_url"
       }
   }
 
@@ -28,12 +30,12 @@ Rails.application.configure do
   config.consider_all_requests_local = true
 
   # Enable/disable caching. By default caching is disabled.
-  if Rails.root.join('tmp/caching-dev.txt').exist?
+  if Rails.root.join("tmp/caching-dev.txt").exist?
     config.action_controller.perform_caching = true
 
     config.cache_store = :memory_store
     config.public_file_server.headers = {
-      'Cache-Control' => 'public, max-age=172800'
+      "Cache-Control" => "public, max-age=172800"
     }
   else
     config.action_controller.perform_caching = false
