@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
@@ -12,7 +14,7 @@ Rails.application.configure do
           access_key_id: ENV["AWS_ACCESS_KEY_ID"],
           secret_access_key: ENV["AWS_SECRET_ACCESS_KEY"],
           path: "image/:id/:filename",
-          url: ':s3_domain_url'
+          url: ":s3_domain_url"
       }
   }
 
@@ -31,7 +33,7 @@ Rails.application.configure do
 
   # Disable serving static files from the `/public` folder by default since
   # Apache or NGINX already handles this.
-  config.public_file_server.enabled = ENV['RAILS_SERVE_STATIC_FILES'].present?
+  config.public_file_server.enabled = ENV["RAILS_SERVE_STATIC_FILES"].present?
 
   # Compress JavaScripts and CSS.
   config.assets.js_compressor = :uglifier
@@ -74,15 +76,15 @@ Rails.application.configure do
 
   # SMTP settings for gmail
   config.action_mailer.smtp_settings = {
-      address: ENV['MAIL_ADDRESS'],
+      address: ENV["MAIL_ADDRESS"],
       port: 587,
-      user_name: ENV['MAIL_USERNAME'],
-      password: ENV['MAIL_PASSWORD'],
-      authentication: 'plain',
+      user_name: ENV["MAIL_USERNAME"],
+      password: ENV["MAIL_PASSWORD"],
+      authentication: "plain",
       enable_starttls_auto: true
   }
 
-  config.action_mailer.default_url_options = {host: 'kudos.kabisa.io'}
+  config.action_mailer.default_url_options = { host: "kudos.kabisa.io" }
 
   # Ignore bad email addresses and do not raise email delivery errors.
   # Set this to true and configure the email server for immediate delivery to raise delivery errors.
