@@ -1,4 +1,6 @@
+# frozen_string_literal: true
+
 begin
-  fcm_server_key = ENV['FCM_SERVER_KEY']
+  fcm_server_key = ENV["FCM_SERVER_KEY"]
   FCM_SERVICE = fcm_server_key.present? ? FCM.new(fcm_server_key) : nil
 end
