@@ -1,1 +1,3 @@
-Rails.application.routes.default_url_options[:host] = ENV['ROOT_URL']
+# frozen_string_literal: true
+
+Rails.application.routes.default_url_options[:host] = ENV["ROOT_URL"]
