@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Be sure to restart your server when you modify this file.
 
-Rails.application.config.session_store :cookie_store, key: '_kudo_o_matic_session'
+Rails.application.config.session_store :cookie_store, key: "_kudo_o_matic_session"
