@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 class RemoveEmailIndexFromUser < ActiveRecord::Migration[5.0]
   def change
     begin
       remove_index :users, :email
     rescue
-      puts 'skipped migration'
+      puts "skipped migration"
     end
   end
 end
