@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class CreateDoorkeeperTables < ActiveRecord::Migration[5.0]
   def change
     create_table :oauth_applications do |t|
@@ -5,7 +7,7 @@ class CreateDoorkeeperTables < ActiveRecord::Migration[5.0]
       t.string  :uid,          null: false
       t.string  :secret,       null: false
       t.text    :redirect_uri, null: false
-      t.string  :scopes,       null: false, default: ''
+      t.string  :scopes,       null: false, default: ""
       t.timestamps             null: false
     end
 
