@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 class RemoveUserIdFromTeamInvites < ActiveRecord::Migration[5.0]
   def change
-    TeamInvite.find_each {|t| t.update(email: User.find(t.user_id).email)}
+    TeamInvite.find_each { |t| t.update(email: User.find(t.user_id).email) }
     remove_column :team_invites, :user_id, :fk
   end
 end
