@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -10,246 +12,255 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20181011145946) do
-
+ActiveRecord::Schema.define(version: 2018_10_30_103959) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "activities", force: :cascade do |t|
-    t.text     "name"
-    t.integer  "suggested_amount"
-    t.datetime "created_at",       null: false
-    t.datetime "updated_at",       null: false
+  create_table "activities", id: :serial, force: :cascade do |t|
+    t.integer "suggested_amount"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "balances", force: :cascade do |t|
-    t.string   "name"
-    t.boolean  "current",    default: false
-    t.datetime "created_at",                 null: false
-    t.datetime "updated_at",                 null: false
-    t.integer  "team_id"
-    t.index ["team_id"], name: "index_balances_on_team_id", using: :btree
+  create_table "balances", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.boolean "current", default: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer "team_id"
+    t.index ["team_id"], name: "index_balances_on_team_id"
   end
 
-  create_table "delayed_jobs", force: :cascade do |t|
-    t.integer  "priority",   default: 0, null: false
-    t.integer  "attempts",   default: 0, null: false
-    t.text     "handler",                null: false
-    t.text     "last_error"
+  create_table "delayed_jobs", id: :serial, force: :cascade do |t|
+    t.integer "priority", default: 0, null: false
+    t.integer "attempts", default: 0, null: false
+    t.text "handler", null: false
+    t.text "last_error"
     t.datetime "run_at"
     t.datetime "locked_at"
     t.datetime "failed_at"
-    t.string   "locked_by"
-    t.string   "queue"
+    t.string "locked_by"
+    t.string "queue"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.index ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
+    t.index ["priority", "run_at"], name: "delayed_jobs_priority"
   end
 
-  create_table "exports", force: :cascade do |t|
-    t.string   "uuid"
-    t.integer  "user_id"
-    t.datetime "created_at",       null: false
-    t.datetime "updated_at",       null: false
-    t.string   "zip_file_name"
-    t.string   "zip_content_type"
-    t.integer  "zip_file_size"
+  create_table "exports", id: :serial, force: :cascade do |t|
+    t.string "uuid"
+    t.integer "user_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "zip_file_name"
+    t.string "zip_content_type"
+    t.bigint "zip_file_size"
     t.datetime "zip_updated_at"
-    t.index ["user_id"], name: "index_exports_on_user_id", using: :btree
+    t.index ["user_id"], name: "index_exports_on_user_id"
   end
 
-  create_table "fcm_tokens", force: :cascade do |t|
-    t.string   "token"
-    t.integer  "user_id"
+  create_table "fcm_tokens", id: :serial, force: :cascade do |t|
+    t.string "token"
+    t.integer "user_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.index ["user_id"], name: "index_fcm_tokens_on_user_id", using: :btree
+    t.index ["user_id"], name: "index_fcm_tokens_on_user_id"
   end
 
-  create_table "friendly_id_slugs", force: :cascade do |t|
-    t.string   "slug",                      null: false
-    t.integer  "sluggable_id",              null: false
-    t.string   "sluggable_type", limit: 50
-    t.string   "scope"
+  create_table "friendly_id_slugs", id: :serial, force: :cascade do |t|
+    t.string "slug", null: false
+    t.integer "sluggable_id", null: false
+    t.string "sluggable_type", limit: 50
+    t.string "scope"
     t.datetime "created_at"
-    t.index ["slug", "sluggable_type", "scope"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type_and_scope", unique: true, using: :btree
-    t.index ["slug", "sluggable_type"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type", using: :btree
-    t.index ["sluggable_id"], name: "index_friendly_id_slugs_on_sluggable_id", using: :btree
-    t.index ["sluggable_type"], name: "index_friendly_id_slugs_on_sluggable_type", using: :btree
+    t.index ["slug", "sluggable_type", "scope"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type_and_scope", unique: true
+    t.index ["slug", "sluggable_type"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type"
+    t.index ["sluggable_id"], name: "index_friendly_id_slugs_on_sluggable_id"
+    t.index ["sluggable_type"], name: "index_friendly_id_slugs_on_sluggable_type"
   end
 
-  create_table "goals", force: :cascade do |t|
-    t.string   "name",        limit: 32
-    t.integer  "amount"
-    t.date     "achieved_on"
-    t.datetime "created_at",             null: false
-    t.datetime "updated_at",             null: false
-    t.integer  "balance_id"
-  end
-
-  create_table "guidelines", force: :cascade do |t|
-    t.string   "name"
-    t.integer  "kudos"
+  create_table "goals", id: :serial, force: :cascade do |t|
+    t.string "name", limit: 32
+    t.integer "amount"
+    t.date "achieved_on"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.integer  "team_id"
-    t.index ["team_id"], name: "index_guidelines_on_team_id", using: :btree
+    t.integer "balance_id"
   end
 
-  create_table "oauth_access_grants", force: :cascade do |t|
-    t.integer  "resource_owner_id", null: false
-    t.integer  "application_id",    null: false
-    t.string   "token",             null: false
-    t.integer  "expires_in",        null: false
-    t.text     "redirect_uri",      null: false
-    t.datetime "created_at",        null: false
+  create_table "guidelines", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.integer "kudos"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer "team_id"
+    t.index ["team_id"], name: "index_guidelines_on_team_id"
+  end
+
+  create_table "oauth_access_grants", id: :serial, force: :cascade do |t|
+    t.integer "resource_owner_id", null: false
+    t.integer "application_id", null: false
+    t.string "token", null: false
+    t.integer "expires_in", null: false
+    t.text "redirect_uri", null: false
+    t.datetime "created_at", null: false
     t.datetime "revoked_at"
-    t.string   "scopes"
-    t.index ["application_id"], name: "index_oauth_access_grants_on_application_id", using: :btree
-    t.index ["token"], name: "index_oauth_access_grants_on_token", unique: true, using: :btree
+    t.string "scopes"
+    t.index ["application_id"], name: "index_oauth_access_grants_on_application_id"
+    t.index ["token"], name: "index_oauth_access_grants_on_token", unique: true
   end
 
-  create_table "oauth_access_tokens", force: :cascade do |t|
-    t.integer  "resource_owner_id"
-    t.integer  "application_id"
-    t.string   "token",                               null: false
-    t.string   "refresh_token"
-    t.integer  "expires_in"
+  create_table "oauth_access_tokens", id: :serial, force: :cascade do |t|
+    t.integer "resource_owner_id"
+    t.integer "application_id"
+    t.string "token", null: false
+    t.string "refresh_token"
+    t.integer "expires_in"
     t.datetime "revoked_at"
-    t.datetime "created_at",                          null: false
-    t.string   "scopes"
-    t.string   "previous_refresh_token", default: "", null: false
-    t.index ["application_id"], name: "index_oauth_access_tokens_on_application_id", using: :btree
-    t.index ["refresh_token"], name: "index_oauth_access_tokens_on_refresh_token", unique: true, using: :btree
-    t.index ["resource_owner_id"], name: "index_oauth_access_tokens_on_resource_owner_id", using: :btree
-    t.index ["token"], name: "index_oauth_access_tokens_on_token", unique: true, using: :btree
+    t.datetime "created_at", null: false
+    t.string "scopes"
+    t.string "previous_refresh_token", default: "", null: false
+    t.index ["application_id"], name: "index_oauth_access_tokens_on_application_id"
+    t.index ["refresh_token"], name: "index_oauth_access_tokens_on_refresh_token", unique: true
+    t.index ["resource_owner_id"], name: "index_oauth_access_tokens_on_resource_owner_id"
+    t.index ["token"], name: "index_oauth_access_tokens_on_token", unique: true
   end
 
-  create_table "oauth_applications", force: :cascade do |t|
-    t.string   "name",                      null: false
-    t.string   "uid",                       null: false
-    t.string   "secret",                    null: false
-    t.text     "redirect_uri",              null: false
-    t.string   "scopes",       default: "", null: false
-    t.datetime "created_at",                null: false
-    t.datetime "updated_at",                null: false
-    t.index ["uid"], name: "index_oauth_applications_on_uid", unique: true, using: :btree
+  create_table "oauth_applications", id: :serial, force: :cascade do |t|
+    t.string "name", null: false
+    t.string "uid", null: false
+    t.string "secret", null: false
+    t.text "redirect_uri", null: false
+    t.string "scopes", default: "", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["uid"], name: "index_oauth_applications_on_uid", unique: true
   end
 
-  create_table "team_invites", force: :cascade do |t|
-    t.integer  "team_id"
-    t.datetime "sent_at"
+  create_table "post_receivers", force: :cascade do |t|
+    t.bigint "user_id"
+    t.bigint "post_id"
+    t.index ["post_id"], name: "index_post_receivers_on_post_id"
+    t.index ["user_id"], name: "index_post_receivers_on_user_id"
+  end
+
+  create_table "posts", id: :serial, force: :cascade do |t|
+    t.integer "sender_id"
+    t.integer "activity_id"
+    t.integer "balance_id"
+    t.integer "amount"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "image_file_name"
+    t.string "image_content_type"
+    t.bigint "image_file_size"
+    t.datetime "image_updated_at"
+    t.string "slack_reaction_created_at"
+    t.string "slack_transaction_updated_at"
+    t.integer "slack_kudos_left_on_creation"
+    t.integer "team_id"
+    t.string "message"
+    t.index ["team_id"], name: "index_posts_on_team_id"
+  end
+
+  create_table "team_invites", id: :serial, force: :cascade do |t|
+    t.integer "team_id"
+    t.datetime "sent_at", default: -> { "CURRENT_TIMESTAMP" }
     t.datetime "accepted_at"
     t.datetime "declined_at"
-    t.string   "email"
-    t.index ["email"], name: "index_team_invites_on_email", using: :btree
-    t.index ["team_id"], name: "index_team_invites_on_team_id", using: :btree
+    t.string "email"
+    t.index ["email"], name: "index_team_invites_on_email"
+    t.index ["team_id"], name: "index_team_invites_on_team_id"
   end
 
-  create_table "team_members", force: :cascade do |t|
-    t.integer  "team_id"
-    t.integer  "user_id"
-    t.boolean  "admin"
-    t.datetime "created_at",     null: false
-    t.datetime "updated_at",     null: false
-    t.string   "slack_id"
-    t.string   "slack_username"
-    t.string   "slack_name"
-    t.index ["team_id"], name: "index_team_members_on_team_id", using: :btree
-    t.index ["user_id"], name: "index_team_members_on_user_id", using: :btree
+  create_table "team_members", id: :serial, force: :cascade do |t|
+    t.integer "team_id"
+    t.integer "user_id"
+    t.boolean "admin"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "slack_id"
+    t.string "slack_username"
+    t.string "slack_name"
+    t.index ["team_id"], name: "index_team_members_on_team_id"
+    t.index ["user_id"], name: "index_team_members_on_user_id"
   end
 
-  create_table "teams", force: :cascade do |t|
-    t.string   "name"
-    t.text     "general_info"
-    t.datetime "created_at",             null: false
-    t.datetime "updated_at",             null: false
-    t.string   "logo_file_name"
-    t.string   "logo_content_type"
-    t.integer  "logo_file_size"
+  create_table "teams", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.text "general_info"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "logo_file_name"
+    t.string "logo_content_type"
+    t.bigint "logo_file_size"
     t.datetime "logo_updated_at"
-    t.string   "slug"
-    t.json     "preferences"
-    t.string   "slack_team_id"
-    t.string   "slack_bot_access_token"
-    t.string   "channel_id"
-    t.string   "rss_token"
+    t.string "slug"
+    t.json "preferences"
+    t.string "slack_team_id"
+    t.string "slack_bot_access_token"
+    t.string "channel_id"
+    t.string "rss_token"
   end
 
-  create_table "transactions", force: :cascade do |t|
-    t.integer  "sender_id"
-    t.integer  "receiver_id"
-    t.integer  "activity_id"
-    t.integer  "balance_id"
-    t.integer  "amount"
-    t.datetime "created_at",                   null: false
-    t.datetime "updated_at",                   null: false
-    t.string   "image_file_name"
-    t.string   "image_content_type"
-    t.integer  "image_file_size"
-    t.datetime "image_updated_at"
-    t.string   "slack_reaction_created_at"
-    t.string   "slack_transaction_updated_at"
-    t.integer  "slack_kudos_left_on_creation"
-    t.integer  "team_id"
-    t.index ["team_id"], name: "index_transactions_on_team_id", using: :btree
-  end
-
-  create_table "users", force: :cascade do |t|
-    t.string   "name"
-    t.datetime "created_at",                             null: false
-    t.datetime "updated_at",                             null: false
-    t.string   "email",                  default: "",    null: false
-    t.string   "encrypted_password",     default: "",    null: false
-    t.string   "reset_password_token"
+  create_table "users", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "email", default: "", null: false
+    t.string "encrypted_password", default: "", null: false
+    t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          default: 0,     null: false
+    t.integer "sign_in_count", default: 0, null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.inet     "current_sign_in_ip"
-    t.inet     "last_sign_in_ip"
-    t.string   "confirmation_token"
+    t.inet "current_sign_in_ip"
+    t.inet "last_sign_in_ip"
+    t.string "confirmation_token"
     t.datetime "confirmed_at"
     t.datetime "confirmation_sent_at"
-    t.string   "unconfirmed_email"
-    t.integer  "failed_attempts",        default: 0,     null: false
-    t.string   "unlock_token"
+    t.string "unconfirmed_email"
+    t.integer "failed_attempts", default: 0, null: false
+    t.string "unlock_token"
     t.datetime "locked_at"
-    t.string   "provider"
-    t.string   "uid"
-    t.string   "avatar_url"
-    t.string   "slack_name"
-    t.boolean  "admin",                  default: false
-    t.string   "api_token"
+    t.string "provider"
+    t.string "uid"
+    t.string "avatar_url"
+    t.string "slack_name"
+    t.boolean "admin", default: false
+    t.string "api_token"
     t.datetime "deactivated_at"
-    t.json     "preferences"
-    t.string   "slack_id"
-    t.string   "slack_username"
-    t.boolean  "restricted",             default: false
-    t.boolean  "company_user",           default: false
-    t.index ["confirmation_token"], name: "index_users_on_confirmation_token", unique: true, using: :btree
-    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
-    t.index ["unlock_token"], name: "index_users_on_unlock_token", unique: true, using: :btree
+    t.json "preferences"
+    t.string "slack_id"
+    t.string "slack_username"
+    t.boolean "restricted", default: false
+    t.boolean "company_user", default: false
+    t.index ["confirmation_token"], name: "index_users_on_confirmation_token", unique: true
+    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
+    t.index ["unlock_token"], name: "index_users_on_unlock_token", unique: true
   end
 
-  create_table "votes", force: :cascade do |t|
-    t.string   "votable_type"
-    t.integer  "votable_id"
-    t.string   "voter_type"
-    t.integer  "voter_id"
-    t.boolean  "vote_flag"
-    t.string   "vote_scope"
-    t.integer  "vote_weight"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.index ["votable_id", "votable_type", "vote_scope"], name: "index_votes_on_votable_id_and_votable_type_and_vote_scope", using: :btree
-    t.index ["voter_id", "voter_type", "vote_scope"], name: "index_votes_on_voter_id_and_voter_type_and_vote_scope", using: :btree
+  create_table "votes", id: :serial, force: :cascade do |t|
+    t.string "votable_type"
+    t.integer "votable_id"
+    t.string "voter_type"
+    t.integer "voter_id"
+    t.boolean "vote_flag"
+    t.string "vote_scope"
+    t.integer "vote_weight"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["votable_id", "votable_type", "vote_scope"], name: "index_votes_on_votable_id_and_votable_type_and_vote_scope"
+    t.index ["votable_type", "votable_id"], name: "index_votes_on_votable_type_and_votable_id"
+    t.index ["voter_id", "voter_type", "vote_scope"], name: "index_votes_on_voter_id_and_voter_type_and_vote_scope"
+    t.index ["voter_type", "voter_id"], name: "index_votes_on_voter_type_and_voter_id"
   end
 
   add_foreign_key "exports", "users"
   add_foreign_key "oauth_access_grants", "oauth_applications", column: "application_id"
   add_foreign_key "oauth_access_tokens", "oauth_applications", column: "application_id"
+  add_foreign_key "post_receivers", "posts"
+  add_foreign_key "post_receivers", "users"
   add_foreign_key "team_invites", "teams"
   add_foreign_key "team_members", "teams"
   add_foreign_key "team_members", "users"
